pyproject.toml
[project]
name = "gitmind-ai-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.1.1",
    "faiss-cpu>=1.12.0",
    "gitpython>=3.1.45",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.79",
    "langchain-google-genai>=2.1.12",
    "langchain-groq>=0.3.8",
    "langgraph>=0.6.10",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.1",
    "streamlit>=1.50.0",
]
